// packages/collectors/code/ts-analyzer/src/analysis/types.ts
export type EntityId = string

export interface EntityLogLocation {
  logStart: number
  logEnd: number
  code: string
}
export interface EntityLocation extends Partial<EntityLogLocation>{
  start: number
  end: number
  lineStart: number
  columnStart: number
  lineEnd: number
  columnEnd: number
  kind: string
}

export interface BaseEntity extends EntityLocation {
  type: CodeInfoType.Entity
  filePath: string
  id: EntityId
  entityType: EntityType
}
export interface NonLocationalBaseEntity {
  type: CodeInfoType.Entity
  filePath: string
  id: EntityId
  entityType: EntityType
}

export interface FileEntity extends NonLocationalBaseEntity {
  entityType: EntityType.File
}
export interface FolderEntity extends NonLocationalBaseEntity {
  entityType: EntityType.Folder
}

export interface FunctionEntity extends BaseEntity {
  entityType: EntityType.Function
  typeString: string
}

export interface IdentifierEntity extends BaseEntity {
  entityType: EntityType.Identifier
}

export interface UnknownEntity extends BaseEntity {
  code: string
  [k: `meta_${string}`]: any
}

export interface ActionEntity {
  type: CodeInfoType.Entity
  entityType: ActionEntityType.Call
  id: EntityId
  executionTime: number
}

export type Entity =
  | FileEntity
  | FolderEntity
  | FunctionEntity
  | IdentifierEntity
  | UnknownEntity
  | BaseEntity
  | ActionEntity

export type FSEntity = Extract<Entity, FileEntity | FolderEntity>
// ...
export enum EntityType {
  File = 'file',
  Folder = 'folder',
  Function = 'function',
  Class = 'class',
  Identifier = 'identifier',
  Variable = 'variable',
  Property = 'property',
  ImportDeclaration = 'import_declaration',
  Unknown = 'unknown',
}
export enum RelationType {
  Import = 'import',
  Use = 'use',
  Parent = 'parent',
  CallEntity = 'call_entity',

}

export interface SimpleRelationship {
  type: CodeInfoType.Relationship
  relationType: RelationType
  from: EntityId
  to: EntityId
  id: string
}

export type LocatedRelationship = SimpleRelationship & EntityLocation

export type Relationship = SimpleRelationship | LocatedRelationship

export enum CodeInfoType {
  Entity = 'entity',
  Relationship = 'relationship',
}